+++ commons-app/src/main/java/eu/unifiedviews/commons/dao/view/PipelineView.java
@@ -53,7 +53,7 @@ public class PipelineView implements Serializable, DataObject {
     }
 
     @Override
-    public Long getId() {
+    public int getId() {
         return id;
     }
